Fix wallpaper upload issue

+++ app.js
@@ -186,26 +186,42 @@ const customize = (customizationData) => {
         if (customizationData.backgroundImage !== null && customizationData.backgroundImage !== undefined) {
           if (!customizationData.backgroundImage.match(/^http/)) {
 
-            // Remove the old wallpaper and rename the new
-            fs.unlink(settings.backgroundImage, (err) => {
-              if (!err || err.code === 'ENOENT') {
-                if (customizationData.backgroundImage.match(' ')) {
-                  let newFilename = customizationData.backgroundImage.split(' ').join('_');
-                  fs.rename(customizationData.backgroundImage, newFilename, (err) => {
-                    if (!err) {
-                      settings.backgroundImage = newFilename;
+            const renamePicture = () => {
+              if (customizationData.backgroundImage.match(' ')) {
+                let newFilename = customizationData.backgroundImage.split(' ').join('_');
+                fs.rename(customizationData.backgroundImage, newFilename, (err) => {
+                  if (!err) {
+                    settings.backgroundImage = newFilename;
+                  } else {
+                    // Avoid errors saying the path does not exist after renaming
+                    if (err.code !== 'ENOENT') {
+                      console.log(`Error renaming background-image : ${err}`);
+                    }
+                  }
+                });
+              } else {
+                settings.backgroundImage = customizationData.backgroundImage;
+              }
+            }
+
+            if (settings.backgroundImage !== undefined) {
+              fs.stat(settings.backgroundImage, (err, stats) => {
+                if (!err) {
+                  // Remove the old wallpaper and rename the new
+                  fs.unlink(settings.backgroundImage, (err) => {
+                    if (!err || err.code === 'ENOENT') {
+                      renamePicture();
                     } else {
-                      // Avoid errors saying the path does not exist after renaming
-                      if (err.code !== 'ENOENT') {
-                        console.log(`Error renaming background-image : ${err}`);
-                      }
+                      console.log(`Error deleting the previous background image : ${err}`);
                     }
                   });
                 } else {
-                  settings.backgroundImage = customizationData.backgroundImage;
+                  renamePicture();
                 }
-              }
-            })
+              });
+            } else {
+              renamePicture();
+            }
           } else {
             settings.backgroundImage = customizationData.backgroundImage;
           }
